+++ deployer/default/deploy/task/deploy.php
@@ -53,10 +53,6 @@ task('deploy', [
     // Read more on https://github.com/sourcebroker/deployer-extended-database#db-backup
     'db:backup',
 
-    // Start buffering http requests. No frontend access possible from now.
-    // Read more on https://github.com/sourcebroker/deployer-extended#buffer-start
-    'buffer:start',
-
     // Truncate caching tables, all cf_* tables
     // Read more on https://github.com/sourcebroker/deployer-extended-database#db-truncate
     'db:truncate',
@@ -74,10 +70,6 @@ task('deploy', [
     // Clear frontend http cache.
     // Read more on https://github.com/sourcebroker/deployer-extended#cache-clear-php-http
     'cache:clear_php_http',
-
-    // Frontend access possible again from now
-    // Read more on https://github.com/sourcebroker/deployer-extended#buffer-stop
-    'buffer:stop',
 
     // Standard deployer task.
     'deploy:unlock',